+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Issue with threads trying to write to a non thread safe list causing install issues
 - Creating vanilla packs or importing not checking for account first
 - Issue with collapsing things without an account causing an NPE [#770]
+- Issue with importing from MultiMC causing issues with post/pre launch commands/wrapper [#765]
 
 ### Misc
 - Consolodate GSON configs